+++ actix-http-test/CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* `TestServer::stop` is now async and will wait for the server and system to shutdown. [#2442]
 * Update `actix-server` to `2.0.0-beta.9`. [#2442]
 * Minimum supported Rust version (MSRV) is now 1.52.